requirements.txt
openai>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
